Planned the test cases JUnit rules / stubbing warnings
Work in progress for #384
